go: make main.go more easily testable
We have reduced the number of branches that can occur within
that file, by making the Client more abstract.

As a bonus, now client just holds configuration and all the
context can be local to the Download (or Upload) method.

If this was C++, Client::Download would be a const method.