ndt7_test.go: remove context timeout
Adding a timeout is like calling for flacky tests. No timeout unless
we want to test the timeout. If there's an issue, it would run
forever and travis will kill the test.

Doing this because I've just noticed a travis failure that has been
caused precisely by this event happening.