Avoid publicly exposing an interface
Long time ago @evfirerob mentioned to me that in Go it's more
idiomatic for the consumer to enforce interfaces if needed.

However, my initial stab at this branch was exposing the
HTTPRequestor interface that I actually initially created
to simplify testing and was not meant to be shared.

This knowledge about interfaces resurfaced today when I was
re-reading my diff, and I realized it's probably better to
expose an `*http.Client` rather than an interface.

This simplification does not seem to complicate much more
testing, and rather allows to hide some details of the way
in which we do testing, so I think it's an improvement.