mlabns: make implementation more flexible (#13)
* mlabns: make implementation more flexible

To this end, expose implementation details that allow to override
the way in which a request is created and serviced.

This is mainly driven by the following OONI needs. When a OONI client
is behind a proxy (e.g. Tor's SOCKS5 proxy), it needs to adjust its
behaviour to still be able to perform ndt7 tests correctly.

To this end, it needs:

1. to contact mlabns using the proxy. In OONI we honour the HTTP_PROXY
environment variable but we also have more specific proxy settings that
are configured through a specialized http.Client. To allow for OONI to
do the latter, this PR exposes the HTTPRequestor publicly, so that we
can override the default with our specialized client. (I can hear an
argument that it's possible to do all of this by just using the proxy
environment variable; yet, we use the specialized client also for
other functionality, including (1) measuring and logging everything
that happens, because we're concerned about censorship and we want
therefore to log what happens at low level; (2) possibly overwriting
DNS resolution to use DoH or alternative name servers. As such we
actually need to have HTTPRequestory public and I don't see how we
can get away with just the HTTP_PROXY environment variable).

2. to modify the mlabns http.Request before it's issued so that it
can append the proper query string arguments. To this end, it needs
to replace the default HTTPRequestMaker with one that calls the
original and then changes the query.

My initial solution to solve the above problems have been vendoring
mlabns, but that's sub-optimal and I'd rather have a more flexible
implementation here. I can also see how others may want to have this
similar kind of flexibility for their packages and/or tests.

See also https://github.com/ooni/probe-engine/commit/dfc0e04dce48ad332e9c0e44ccfd31eb3594f928

* Avoid publicly exposing an interface

Long time ago @evfirerob mentioned to me that in Go it's more
idiomatic for the consumer to enforce interfaces if needed.

However, my initial stab at this branch was exposing the
HTTPRequestor interface that I actually initially created
to simplify testing and was not meant to be shared.

This knowledge about interfaces resurfaced today when I was
re-reading my diff, and I realized it's probably better to
expose an `*http.Client` rather than an interface.

This simplification does not seem to complicate much more
testing, and rather allows to hide some details of the way
in which we do testing, so I think it's an improvement.

* mlabns.go: sort initialization